Moved files and renamed EditableField to EditableDescription
