+++ app/assets/stylesheets/new_dashboard/editable_fields/editable-tags.css.scss
