+++ lib/assets/javascripts/cartodb/new_dashboard/datasets/datasets_item.js
@@ -4,8 +4,8 @@ var Utils = require('cdb.Utils');
 var navigateThroughRouter = require('../../new_common/view_helpers/navigate_through_router');
 var pluralizeString = require('../../new_common/view_helpers/pluralize_string');
 var LikesView = require('../../new_common/views/likes/view');
-var EditableField = require('../../new_common/editable_field');
-var EditableTags = require('../../new_common/editable_tags');
+var EditableDescription = require('../../new_dashboard/editable_fields/editable_description');
+var EditableTags = require('../../new_dashboard/editable_fields/editable_tags');
 
 /**
  * View representing an item in the list under datasets route.
@@ -92,10 +92,9 @@ module.exports = cdb.core.View.extend({
 
   _renderDescription: function() {
     var isOwner = this.model.permission.isOwner(this.user);
-    var view = new EditableField({
+    var view = new EditableDescription({
       el: this.$('.DatasetsList-itemDescription'),
       model: this.model,
-      fieldName: 'description',
       maxLength: 200,
       editable: isOwner
     });
