+++ lib/assets/javascripts/cartodb/new_dashboard/editable_fields/editable_description.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js');
 
 /**
- *  Editable text field
+ *  Editable description
  */
 module.exports = cdb.core.View.extend({
 
@@ -18,24 +18,22 @@ module.exports = cdb.core.View.extend({
   },
 
   initialize: function() {
-    this.template = cdb.templates.getTemplate('new_common/views/editable_field');
-    this.fieldName = this.options.fieldName;
+    this.template = cdb.templates.getTemplate('new_dashboard/editable_fields/editable_description');
     this.maxLength = this.options.maxLength;
     this.editable = this.options.editable;
     _.defaults(this.options, this.defaults);
 
     // Backbone's className won't work here because we are providing an el
-    this.$el.addClass('EditableField');
+    this.$el.addClass('EditableDescription');
   },
 
   render: function() {
-    var safeHTML = cdb.core.sanitize.html(markdown.toHTML(this.model.get(this.fieldName) || ''))
+    var safeHTML = cdb.core.sanitize.html(markdown.toHTML(this.model.get('description') || ''))
     var value = {
       safeHTML: safeHTML,
       clean: cdb.Utils.stripHTML(safeHTML)
     };
     this.$el.html(this.template({
-      name: this.fieldName,
       value: value,
       editable: this.editable
     }));
@@ -47,7 +45,7 @@ module.exports = cdb.core.View.extend({
     this.killEvent(ev);
 
     this.$el.addClass('is-editing');
-    this.$('.js-field-input').focus();
+    this.$('.js-field-input').val('').focus();
   },
 
   _keyPressed: function(ev) {
@@ -71,19 +69,18 @@ module.exports = cdb.core.View.extend({
 
   _addNewLine: function() {
     this.$('.js-field-input').val(this.$('.js-field-input').val() + "\n");
-    var attributes = {};
-    attributes[this.fieldName] = this.$('.js-field-input').val();
+    //TODO: Scroll to bottom of textarea
   },
 
   _save: function() {
-    var attributes = {};
-    attributes[this.fieldName] = this.$('.js-field-input').val();
+    var attributes = {
+      description: this.$('.js-field-input').val()
+    };
     this.model.save(attributes);
     this.render();
   },
 
   _cancelEditing: function() {
     this.$el.removeClass('is-editing');
-    this.$('.js-field-input').val(this.model.get(this.fieldName));
   }
 });
