+++ lib/assets/javascripts/cartodb/new_dashboard/editable_fields/editable_description.jst.ejs
@@ -2,7 +2,7 @@
   <div class="DefaultDescription" title="<%- value.clean %>"><%= value.safeHTML %></div>
 <% } else { %>
   <% if (editable) { %>
-    <button class="js-add-btn">Add <%= name %>...</button>
+    <button class="js-add-btn">Add description...</button>
     <textarea class="js-field-input"></textarea>
   <% } else { %>
   <div class="DefaultDescription">
