+++ lib/assets/javascripts/cartodb/new_dashboard/editable_fields/editable_tags.js
@@ -35,7 +35,6 @@ module.exports = cdb.core.View.extend({
 
   _edit: function(ev) {
     this.killEvent(ev);
-
     this.$el.addClass('is-editing');
     this.$('.js-field-input').val('').focus();
   },
@@ -43,7 +42,6 @@ module.exports = cdb.core.View.extend({
   _keyPressed: function(ev) {
     var enterPressed = (ev.keyCode == $.ui.keyCode.ENTER);
     var escapePressed = (ev.keyCode == $.ui.keyCode.ESCAPE);
-
     if (enterPressed) {
       this._save();
     } else if (escapePressed) {
@@ -52,18 +50,14 @@ module.exports = cdb.core.View.extend({
   },
 
   _save: function() {
-    var attributes = {
-      tags: this._sanitizedTags()
-    };
-    this.model.save(attributes);
-    this.render();
-  },
-
-  _sanitizedTags: function() {
     var tags = this.$('.js-field-input').val().split(',').map(function(tag){
       return cdb.Utils.stripHTML(tag.trim());
     })
-    return _.uniq(tags);
+    tags = _.uniq(tags);
+    this.model.save({
+      tags: tags
+    });
+    this.render();
   },
 
   _cancelEditing: function(ev) {
