+++ lib/assets/javascripts/cartodb/new_dashboard/editable_fields/editable_tags.jst.ejs
