+++ lib/assets/javascripts/cartodb/new_dashboard/maps/maps_item.js
@@ -4,8 +4,8 @@ var navigateThroughRouter = require('../../new_common/view_helpers/navigate_thro
 var MapviewsGraph = require('../../new_dashboard/mapviews_graph');
 var MapCardPreview = require('../../new_common/views/mapcard_preview');
 var LikesView = require('../../new_common/views/likes/view');
-var EditableField = require('../../new_common/editable_field');
-var EditableTags = require('../../new_common/editable_tags');
+var EditableDescription= require('../../new_dashboard/editable_fields/editable_description');
+var EditableTags = require('../../new_dashboard/editable_fields/editable_tags');
 var Utils = require('cdb.Utils');
 cdb.admin = require('cdb.admin');
 
@@ -72,7 +72,7 @@ module.exports = cdb.core.View.extend({
 
   _renderDescription: function() {
     var isOwner = this.model.permission.isOwner(this.user);
-    var view = new EditableField({
+    var view = new EditableDescription({
       el: this.$('.MapCard-desc'),
       model: this.model,
       fieldName: 'description',
