+++ lib/assets/test/spec/cartodb/new_dashboard/datasets/datasets_item.spec.js
@@ -125,7 +125,7 @@ describe('new_dashboard/datasets/datasets_item', function() {
   it('should render an editable field with the description', function() {
     this.vis.set('description', 'my desc');
     this.renderView();
-    expect(this.view.$('.DatasetsList-itemDescription.EditableField').length).toBe(1);
+    expect(this.view.$('.DatasetsList-itemDescription.EditableDescription').length).toBe(1);
     expect(this.view.$('.DefaultDescription').length).toBe(1);
     expect(this.view.$('.DefaultDescription')[0].textContent).toEqual('my desc');
   });
