+++ lib/assets/test/spec/cartodb/new_dashboard/editable_fields/editable_description.spec.js
@@ -1,4 +1,4 @@
-var EditableField = require('../../../../javascripts/cartodb/new_common/editable_field');
+var EditableDescription = require('../../../../../javascripts/cartodb/new_dashboard/editable_fields/editable_description');
 
 function keyPressEvent(key, metaKey) {
   var event = jQuery.Event("keydown");
@@ -23,24 +23,23 @@ function type($textarea, word) {
 
 var model, view;
 
-describe('new_common/editable_field', function() {
+describe('new_dashboard/editable_fields/editable_description', function() {
 
   beforeEach(function() {
     model = new Backbone.Model({
       description: ''
     });
-    view = new EditableField({
-      model: model,
-      fieldName: 'description'
+    view = new EditableDescription({
+      model: model
     });
     spyOn(Backbone, 'sync');
     spyOn(model, 'save').and.callThrough();
   });
 
-  it('should add the "EditableField" class to the element', function() {
+  it('should add the "EditableDescription" class to the element', function() {
     view.render();
 
-    expect(view.el.classList).toContain('EditableField');
+    expect(view.el.classList).toContain('EditableDescription');
   });
 
   it('should render the field when it has some value', function() {
@@ -56,7 +55,7 @@ describe('new_common/editable_field', function() {
   describe('if the field is editable', function(){
 
     beforeEach(function() {
-      view = new EditableField({
+      view = new EditableDescription({
         model: model,
         fieldName: 'description',
         editable: true
@@ -102,12 +101,10 @@ describe('new_common/editable_field', function() {
       type(view.$('.js-field-input'), "wadus");
       view.$('.js-field-input').trigger(keyPressEvent($.ui.keyCode.ESCAPE));
 
-      expect(view.$('.js-field-input').val()).toEqual('');
       expect(view.el.classList).not.toContain('is-editing');
 
       // Model has not been updated
       expect(model.save).not.toHaveBeenCalled();
-      expect(model.get('description')).toEqual('');
     });
 
     it('should cancel edition on blur', function() {
@@ -122,7 +119,6 @@ describe('new_common/editable_field', function() {
       type(view.$('.js-field-input'), "wadus");
       view.$('.js-field-input').trigger(jQuery.Event("blur"));
 
-      expect(view.$('.js-field-input').val()).toEqual('');
       expect(view.el.classList).not.toContain('is-editing');
     });
 
@@ -159,7 +155,7 @@ describe('new_common/editable_field', function() {
   describe('if the field is NOT editable', function(){
 
     beforeEach(function() {
-      view = new EditableField({
+      view = new EditableDescription({
         model: model,
         fieldName: 'description',
         editable: false
