+++ lib/assets/test/spec/cartodb/new_dashboard/editable_fields/editable_tags.spec.js
@@ -1,4 +1,4 @@
-var EditableTags = require('../../../../javascripts/cartodb/new_common/editable_tags');
+var EditableTags = require('../../../../../javascripts/cartodb/new_dashboard/editable_fields/editable_tags');
 
 function keyPressEvent(key, metaKey) {
   var event = jQuery.Event("keydown");
@@ -21,7 +21,7 @@ function type($textarea, word) {
   }
 }
 
-describe('new_common/editable_tags', function() {
+describe('new_dashboard/editable_fields/editable_tags', function() {
 
   var model, view, router;
