+++ lib/assets/test/spec/cartodb/new_dashboard/maps/maps_item.spec.js
@@ -134,7 +134,7 @@ describe('new_dashboard/maps/maps_item', function() {
   it('should render an editable field with the description', function() {
     this.vis.set('description', 'my desc');
     this.renderView();
-    expect(this.view.$('.MapCard-desc.EditableField').length).toBe(1);
+    expect(this.view.$('.MapCard-desc.EditableDescription').length).toBe(1);
     expect(this.view.$('.DefaultDescription').length).toBe(1);
     expect(this.view.$('.DefaultDescription')[0].textContent).toEqual('my desc');
   });
